tech(api): move test
