+++ api/tests/evaluation/unit/application/assessments/assessment-controller_test.js
@@ -11,6 +11,7 @@ import { DomainTransaction } from '../../../../../src/shared/domain/DomainTransa
 import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
 import { featureToggles } from '../../../../../src/shared/infrastructure/feature-toggles/index.js';
 import { expect } from '../../../../test-helper.js';
+import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder.js';
 import { hFake } from '../../../../tooling/mocks/hapi.mock.js';
 
 describe('Evaluation | Unit | Application | assessment-controller', function () {
@@ -292,6 +293,81 @@ describe('Evaluation | Unit | Application | assessment-controller', function () 
         // then
         expect(assessmentRepositoryStub.save).to.have.been.calledWithExactly({ assessment: expected });
       });
+    });
+  });
+
+  describe('#findCompetenceEvaluations', function () {
+    it('should return the competence evaluations', async function () {
+      // given
+      const userId = 123;
+      const assessmentId = 456;
+      const competenceEvaluation1 = domainBuilder.buildCompetenceEvaluation({ assessmentId, userId });
+      const competenceEvaluation2 = domainBuilder.buildCompetenceEvaluation({ assessmentId, userId });
+      sinon
+        .stub(evaluationUsecases, 'findCompetenceEvaluationsByAssessment')
+        .withArgs({ assessmentId, userId })
+        .resolves([competenceEvaluation1, competenceEvaluation2]);
+      const request = {
+        auth: { credentials: { userId } },
+        params: {
+          id: assessmentId,
+        },
+      };
+
+      // when
+      const result = await assessmentController.findCompetenceEvaluations(request, hFake);
+
+      // then
+      expect(result.data).to.be.deep.equal([
+        {
+          type: 'competence-evaluations',
+          id: competenceEvaluation1.id.toString(),
+          attributes: {
+            'competence-id': competenceEvaluation1.competenceId,
+            'user-id': competenceEvaluation1.userId,
+            'created-at': competenceEvaluation1.createdAt,
+            'updated-at': competenceEvaluation1.updatedAt,
+            status: competenceEvaluation1.status,
+          },
+          relationships: {
+            assessment: {
+              data: {
+                id: assessmentId.toString(),
+                type: 'assessments',
+              },
+            },
+            scorecard: {
+              links: {
+                related: `/api/scorecards/${userId}_${competenceEvaluation1.competenceId}`,
+              },
+            },
+          },
+        },
+        {
+          type: 'competence-evaluations',
+          id: competenceEvaluation2.id.toString(),
+          attributes: {
+            'competence-id': competenceEvaluation2.competenceId,
+            'user-id': competenceEvaluation2.userId,
+            'created-at': competenceEvaluation2.createdAt,
+            'updated-at': competenceEvaluation2.updatedAt,
+            status: competenceEvaluation2.status,
+          },
+          relationships: {
+            assessment: {
+              data: {
+                id: assessmentId.toString(),
+                type: 'assessments',
+              },
+            },
+            scorecard: {
+              links: {
+                related: `/api/scorecards/${userId}_${competenceEvaluation2.competenceId}`,
+              },
+            },
+          },
+        },
+      ]);
     });
   });
 });
